+++ app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/transfer/TransferUIModel.kt
@@ -6,5 +6,5 @@ data class TransferUIModel(
     val type  : String? = null,
     val sender_id : String? = null,
     val receiver_id : String? = null,
-    val amount : Double? = null
+    val amount : Long? = null
 )
